Use new snapshots instead of old presenters
