+++ jason/src/lib.rs
@@ -15,6 +15,7 @@ pub mod api;
 pub mod media;
 pub mod peer;
 pub mod rpc;
+mod snapshots;
 
 // When the `console_error_panic_hook` feature is enabled, we can call the
 // `set_panic_hook` function at least once during initialization, and then
