--- proto/client-api/src/presenters/mod.rs
@@ -1,7 +1,0 @@
-pub mod peer_presenter;
-pub mod room_presenter;
-pub mod track_presenter;
-
-pub use peer_presenter::PeerPresenter;
-pub use room_presenter::RoomPresenter;
-pub use track_presenter::TrackPresenter;
