+++ proto/client-api/src/snapshots/track.rs
@@ -5,13 +5,14 @@ use medea_reactive::{
     collections::{vec::ObservableVec, ObservableHashMap},
     Observable, ObservableCell,
 };
-use serde::{Serialize, Deserialize};
+use serde::{Deserialize, Serialize};
 
 use crate::{
     Direction, EventHandler, IceCandidate, IceServer, MediaType, PeerId, Track,
     TrackId, TrackPatch,
 };
 
+#[derive(Debug)]
 pub struct TrackSnapshot {
     pub id: TrackId,
     pub is_muted: bool,
@@ -20,7 +21,12 @@ pub struct TrackSnapshot {
 }
 
 pub trait TrackSnapshotAccessor {
-    fn new(id: TrackId, is_muted: bool, direction: Direction, media_type: MediaType) -> Self;
+    fn new(
+        id: TrackId,
+        is_muted: bool,
+        direction: Direction,
+        media_type: MediaType,
+    ) -> Self;
 
     fn update(&mut self, patch: TrackPatch) {
         if let Some(is_muted) = patch.is_muted {
@@ -40,7 +46,12 @@ pub trait TrackSnapshotAccessor {
 }
 
 impl TrackSnapshotAccessor for TrackSnapshot {
-    fn new(id: TrackId, is_muted: bool, direction: Direction, media_type: MediaType) -> Self {
+    fn new(
+        id: TrackId,
+        is_muted: bool,
+        direction: Direction,
+        media_type: MediaType,
+    ) -> Self {
         Self {
             id,
             is_muted,
@@ -66,46 +77,6 @@ impl TrackSnapshotAccessor for TrackSnapshot {
     }
 
     fn get_id(&self) -> TrackId {
-        self.id
-    }
-}
-
-
-
-
-
-#[derive(Debug)]
-pub struct TrackPresenter {
-    pub(super) id: TrackId,
-    pub(super) is_muted: ObservableCell<bool>,
-    pub(super) direction: Direction,
-    pub(super) media_type: MediaType,
-}
-
-impl TrackPresenter {
-    pub fn update(&mut self, patch: TrackPatch) {
-        if let Some(is_muted) = patch.is_muted {
-            self.is_muted.set(is_muted);
-        }
-    }
-
-    pub fn on_track_update(&self) -> impl Stream<Item = bool> {
-        self.is_muted.subscribe()
-    }
-
-    pub fn get_direction(&self) -> &Direction {
-        &self.direction
-    }
-
-    pub fn get_media_type(&self) -> &MediaType {
-        &self.media_type
-    }
-
-    pub fn get_is_muted(&self) -> bool {
-        self.is_muted.get()
-    }
-
-    pub fn get_id(&self) -> TrackId {
         self.id
     }
 }
